+++ app/pb_kits/playbook/pb_text_input/docs/_text_input_default.jsx
@@ -20,6 +20,9 @@ class TextInputDefault extends React.Component {
     return (
       <div>
         <TextInput
+            aria={{ something: 'hello' }}
+            data={{ say: 'hi', yell: 'go' }}
+            id="unique-id"
             label="First Name"
             placeholder="Enter first name"
             value="Timothy Wenhold"