+++ spec/pb_kits/playbook/kits/label_pill_spec.rb
@@ -13,9 +13,8 @@ RSpec.describe Playbook::PbLabelPill::LabelPill do
                       .with_values("neutral", "success", "warning", "error", "info", "primary") }
   describe "#classname" do
     it "returns namespaced class name", :aggregate_failures do
-      expect(subject.new({}).classname).to eq "pb_label_pill_kit_neutral"
-      expect(subject.new(variant: "success").classname).to eq "pb_label_pill_kit_success"
-      expect(subject.new(classname: "additional_class").classname).to eq "pb_label_pill_kit_neutral additional_class"
+      expect(subject.new({}).classname).to eq "pb_label_pill_kit"
+      expect(subject.new(classname: "additional_class").classname).to eq "pb_label_pill_kit additional_class"
     end
   end
 end